Simplify permission check (#10)
The permission handling from the original version broke a lot of plugins, because for example stores could not be written to even though this was ok (for example in the statistics plugin). Additionally other plugins could no longer be configured, although it might be reasonable to do so, for example for CI plugins.