Fix application of 'ignore lfs' option
The 'ignore lfs' option has been overwritten by the global configuration
and therefore has not been applied to the mirror command correctly.
This was the case, because the flag has been treated as a filter flag
and therefore was ignored when 'overwrite global filter configuration'
has not been selected. Now, the option is always taken from the local configuration.

Co-authored-by: Eduard Heimbuch<eduard.heimbuch@cloudogu.com>
Committed-by: René Pfeuffer<rene.pfeuffer@cloudogu.com>